+++ frontend/src/components/common/async-loading-boundary/custom-async-loading-boundary.spec.tsx
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { mockI18n } from '../../markdown-renderer/test-utils/mock-i18n'
+import { mockI18n } from '../../../test-utils/mock-i18n'
 import { CustomAsyncLoadingBoundary } from './custom-async-loading-boundary'
 import { render } from '@testing-library/react'
