+++ frontend/src/components/common/note-loading-boundary/note-loading-boundary.spec.tsx
@@ -9,9 +9,9 @@ import type { Note } from '../../../api/notes/types'
 import * as LoadingScreenModule from '../../../components/application-loader/loading-screen/loading-screen'
 import * as useSingleStringUrlParameterModule from '../../../hooks/common/use-single-string-url-parameter'
 import * as setNoteDataFromServerModule from '../../../redux/note-details/methods'
+import { mockI18n } from '../../../test-utils/mock-i18n'
 import { testId } from '../../../utils/test-id'
 import * as CommonErrorPageModule from '../../error-pages/common-error-page'
-import { mockI18n } from '../../markdown-renderer/test-utils/mock-i18n'
 import * as CreateNonExistingNoteHintModule from './create-non-existing-note-hint'
 import { NoteLoadingBoundary } from './note-loading-boundary'
 import { render, screen } from '@testing-library/react'
