+++ frontend/src/components/editor-page/document-bar/aliases/aliases-add-form.spec.tsx
@@ -6,8 +6,8 @@
 import * as AliasModule from '../../../../api/alias'
 import * as NoteDetailsReduxModule from '../../../../redux/note-details/methods'
 import type { NoteDetails } from '../../../../redux/note-details/types/note-details'
+import { mockI18n } from '../../../../test-utils/mock-i18n'
 import { mockNoteOwnership } from '../../../../test-utils/note-ownership'
-import { mockI18n } from '../../../markdown-renderer/test-utils/mock-i18n'
 import * as useUiNotificationsModule from '../../../notifications/ui-notification-boundary'
 import { AliasesAddForm } from './aliases-add-form'
 import { render, act, screen } from '@testing-library/react'
