+++ frontend/src/components/editor-page/document-bar/aliases/aliases-list-entry.spec.tsx
@@ -6,8 +6,8 @@
 import * as AliasModule from '../../../../api/alias'
 import type { Alias } from '../../../../api/alias/types'
 import * as NoteDetailsReduxModule from '../../../../redux/note-details/methods'
+import { mockI18n } from '../../../../test-utils/mock-i18n'
 import { mockNoteOwnership } from '../../../../test-utils/note-ownership'
-import { mockI18n } from '../../../markdown-renderer/test-utils/mock-i18n'
 import * as useUiNotificationsModule from '../../../notifications/ui-notification-boundary'
 import { AliasesListEntry } from './aliases-list-entry'
 import { render, act, screen } from '@testing-library/react'
