+++ frontend/src/components/editor-page/editor-pane/linter/single-line-regex-linter.spec.ts
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { mockI18n } from '../../../markdown-renderer/test-utils/mock-i18n'
+import { mockI18n } from '../../../../test-utils/mock-i18n'
 import { SingleLineRegexLinter } from './single-line-regex-linter'
 import type { Diagnostic } from '@codemirror/lint'
 import type { EditorState, Text } from '@codemirror/state'
