+++ frontend/src/components/layout/settings-dialog/utils/on-off-button-group.spec.tsx
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { mockI18n } from '../../../markdown-renderer/test-utils/mock-i18n'
+import { mockI18n } from '../../../../test-utils/mock-i18n'
 import { OnOffButtonGroup } from './on-off-button-group'
 import { act, render, screen } from '@testing-library/react'
