+++ frontend/src/extensions/extra-integrations/highlighted-code-fence/highlighted-code-markdown-extension.spec.tsx
@@ -3,8 +3,8 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { mockI18n } from '../../../components/markdown-renderer/test-utils/mock-i18n'
 import { TestMarkdownRenderer } from '../../../components/markdown-renderer/test-utils/test-markdown-renderer'
+import { mockI18n } from '../../../test-utils/mock-i18n'
 import type { HighlightedCodeProps } from './highlighted-code'
 import * as HighlightedCodeModule from './highlighted-code'
 import { HighlightedCodeMarkdownExtension } from './highlighted-code-markdown-extension'
